Add codegen option instead of derive
doesn't cache the output of the proc macro, and using the derive form of
this code can cause slowdowns in `cargo check` and `rust-analyzer`.

As an alternative, this commit introduces an option to generate the
`PrimeField` implementation in a `build.rs` script, to avoid the
slowdown.
