+++ Cargo.toml
@@ -1,3 +1,10 @@
+[workspace]
+members = [
+    ".",
+    "ff_codegen",
+    "ff_derive",
+]
+
 [package]
 name = "ff"
 version = "0.12.0"
